http_static: misc bug fixes
The request vector generated by hss_ts_rx_callback() must be NULL
terminated.

The hss_main_t use_ptr_thresh member must be a u64 since
unformat_memory_size() expects it. Otherwise, the adjacent u8
enable_url_handlers may have an accident.

Type: fix

+++ src/plugins/http_static/http_static.h
@@ -145,7 +145,7 @@ typedef struct
   /** The bind URI, defaults to tcp://0.0.0.0/80 */
   u8 *uri;
   /** Threshold for switching to ptr data in http msgs */
-  u32 use_ptr_thresh;
+  u64 use_ptr_thresh;
   /** Enable the use of builtinurls */
   u8 enable_url_handlers;
   /** Max cache size before LRU occurs */
